Bump Electron. Yeet the Redux! And all the terrible things of JS world
No words to describe this madness.
Silly me who wanted to actually use this on Mac and Linux.
Docs say it's going to be easier with the newer version? Call me optimistic for not noticing the trap.
The newer version comes with our change in philosophy to use Jekson rather than Jacson, so configs need to change, and Jekson doesn't like your Mamoon lib version so you need to change it.
APIs are also different because we don't like the old way anymore. We are JS, so we add features, not maintain. Why maintain when you can just delete?
And that is actually a library created 4 years ago to be the standard for the modern era. 30 commits were enough. Need more? Use npm fund to give me money.
Hate you, JS world!